refactor after remove old version
