+++ src/DataTable/index.ts
@@ -1,0 +1,14 @@
+export * from "./Components"
+export * from "./Context"
+export * from "./hooks"
+export * from "./types"
+
+export { default as DataTable } from "./DataTable"
+export { default as DataTableBody } from "./DataTableBody"
+export { default as DataTableHead } from "./DataTableHead"
+
+export type { DataTableProps } from "./DataTable"
+export type { DataTableBodyProps } from "./DataTableBody"
+export type { DataTableHeadProps } from "./DataTableHead"
+
+
