+++ src/FilterContainer/index.ts
@@ -1,0 +1,2 @@
+export { default as FilterContainer } from "./FilterContainer"
+export type { FilterContainerProps } from "./FilterContainer"
