--- src/NewDataTable/hooks/index.ts
@@ -1,7 +1,0 @@
-export { default as useSorting } from "./useSorting"
-export { default as useFiltering } from "./useFiltering"
-export { default as useSelectingRows } from "./useSelectingRows"
-
-export * from "./useSorting"
-export * from "./useFiltering"
-export * from "./useSelectingRows"
