--- src/NewDataTable/index.ts
@@ -1,3 +1,0 @@
-export { default as DataTableBody } from "./DataTableBody"
-export { default as NewDataTable } from "./NewDataTable"
-export * from "./types"
