--- src/NewDataTable/types.ts
@@ -1,73 +1,0 @@
-import { HeadCellProps } from '../Components/HeadCell/HeadCell';
-import { CellProps, RowProps, TableProps } from "@grossb/react-table"
-import { ReactNode } from "react"
-
-export type DataTableCell = {
-  id: string | number,
-  value: any,
-  render?: (value: any) => JSX.Element,
-}
-
-export type DataTableHeadCell = DataTableCell & {
-  config?: HeadCellProps
-}
-
-export type DataTableBodyCell = DataTableCell & {
-  config?: CellProps,
-}
-
-export type DataTableRow = {
-  id: string | number,
-  config?: RowProps,
-  render?: (props: any) => JSX.Element,
-}
-
-export type DataTableHeadRow = DataTableRow & {
-  cells?: Array<DataTableHeadCell>,
-}
-
-export type DataTableBodyRow = DataTableRow & {
-  cells?: Array<DataTableBodyCell>,
-}
-
-export type ProcessedDataTableProps = {
-  headRows?: Array<DataTableHeadRow>,
-  bodyRows?: Array<DataTableBodyRow>
-  tableProps?: TableProps,
-  onRowClick?: (event: any, row: DataTableBodyRow) => void
-  selectable?: boolean
-  disableSetCheckboxAfterRowClick?: boolean
-}
-
-export type DataTableApi = {
-  onSelect?: (row: DataTableBodyRow, isSelected: boolean) => void,
-  onBodyCheckboxClick?: (row: DataTableBodyRow) => void,
-  onRowClick?: (event: any, row: DataTableBodyRow) => void,
-  onCheckboxClick?: (event: any, row: DataTableBodyRow) => void,
-}
-
-export type ColumnFilter = {
-  search?: string,
-  [key: string]: any,
-}
-
-export type DataTableColumn = {
-  id?: string,
-  dataField?: string,
-  header?: ReactNode,
-  valueGetter?: (value: DataRow) => ReactNode | JSX.Element,
-  rowComparer?: (first: DataRow, second: DataRow) => -1 | 0 | 1
-  rowFilter?: (row: DataRow, filter?: ColumnFilter) => boolean
-  filterComponent?: ReactNode
-  filterable?: boolean,
-  sortable?: boolean,
-  headCellProps?: ((column: DataTableColumn) => HeadCellProps) | HeadCellProps
-  bodyCellProps?: ((column: DataTableColumn) => CellProps) | CellProps
-}
-
-export type DataRow = {
-  id: number | string,
-  [key: string]: any,
-}
-
-export type RowPropsWithoutChildren = Omit<RowProps, "children">
